+++ bridgetown-core/test/test_regenerator.rb
@@ -5,7 +5,7 @@ require "helper"
 class TestRegenerator < BridgetownUnitTest
   context "The site regenerator" do
     setup do
-      FileUtils.rm_rf(source_dir(".bridgetown-metadata"))
+      FileUtils.rm_rf(site_root_dir(".bridgetown-metadata"))
 
       @site = fixture_site(
         "collections" => {
@@ -93,7 +93,7 @@ class TestRegenerator < BridgetownUnitTest
 
   context "The site regenerator" do
     setup do
-      FileUtils.rm_rf(source_dir(".bridgetown-metadata"))
+      FileUtils.rm_rf(site_root_dir(".bridgetown-metadata"))
       @site = fixture_site(
         "incremental" => true
       )
@@ -127,9 +127,10 @@ class TestRegenerator < BridgetownUnitTest
 
   context "The site metadata" do
     setup do
-      FileUtils.rm_rf(source_dir(".bridgetown-metadata"))
+      FileUtils.rm_rf(site_root_dir(".bridgetown-metadata"))
 
       @site = Site.new(Bridgetown.configuration(
+                         "root_dir"    => site_root_dir,
                          "source"      => source_dir,
                          "destination" => dest_dir,
                          "incremental" => true
@@ -162,7 +163,7 @@ class TestRegenerator < BridgetownUnitTest
       @regenerator.clear
       @regenerator.add(@path)
       @regenerator.write_metadata
-      assert File.file?(source_dir(".bridgetown-metadata"))
+      assert File.file?(site_root_dir(".bridgetown-metadata"))
     end
 
     should "read from the metadata file" do
@@ -171,7 +172,7 @@ class TestRegenerator < BridgetownUnitTest
     end
 
     should "read legacy YAML metadata" do
-      metadata_file = source_dir(".bridgetown-metadata")
+      metadata_file = site_root_dir(".bridgetown-metadata")
       @regenerator = Regenerator.new(@site)
 
       File.open(metadata_file, "w") do |f|
@@ -183,7 +184,7 @@ class TestRegenerator < BridgetownUnitTest
     end
 
     should "not crash when reading corrupted marshal file" do
-      metadata_file = source_dir(".bridgetown-metadata")
+      metadata_file = site_root_dir(".bridgetown-metadata")
       File.open(metadata_file, "w") do |file|
         file.puts Marshal.dump(:foo => "bar")[0, 5]
       end
@@ -311,7 +312,7 @@ class TestRegenerator < BridgetownUnitTest
 
   context "when incremental regeneration is disabled" do
     setup do
-      FileUtils.rm_rf(source_dir(".bridgetown-metadata"))
+      FileUtils.rm_rf(site_root_dir(".bridgetown-metadata"))
       @site = Site.new(Bridgetown.configuration(
                          "source"      => source_dir,
                          "destination" => dest_dir,
@@ -324,7 +325,7 @@ class TestRegenerator < BridgetownUnitTest
     end
 
     should "not create .bridgetown-metadata" do
-      refute File.file?(source_dir(".bridgetown-metadata"))
+      refute File.file?(site_root_dir(".bridgetown-metadata"))
     end
   end
 end